Rename semaphore-related items
