+++ include/semaphore_internal.h
