+++ tests/meson.build
@@ -95,7 +95,7 @@ unit_tests = [
     {'name': 'rect', 'deps': []},
     {'name': 'rgb', 'deps': []},
     {'name': 'rwqueue', 'deps': [libmisc_stubs_dep, libshell_stubs_dep]},
-    {'name': 'semaphore', 'deps': [dosbox_dep]},
+    {'name': 'semaphore_internal', 'deps': [dosbox_dep]},
     {'name': 'setup', 'deps': [dosbox_dep]},
     {'name': 'shell_cmds', 'deps': [dosbox_dep], 'extra_cpp': []},
     {'name': 'shell_redirection', 'deps': [dosbox_dep], 'extra_cpp': []},
